Update K/V hashtrees to support key estimation
This commit updates riak_kv_index_hashtree to provide an API around
the new AAE-based key count estimation feature added to the hashtree
module in riak_core.

This provides a mechanism to estimate the number of keys in a given K/V
vnode based on its AAE trees.

This estimate is used to enable an improvement in the operation of
AAE-based fullsync replication in Riak Enterprise.

Authored-by: Mikael Lixenstrand <mikael.lixenstrand@erlang-solutions.com>
Rebased-by:  Joseph Blomstedt <joe@basho.com>