Expose delete_mode to riak clients
This change exposes the delete_mode parameter to
riak clients, so it can be controlled on individual
DELETE operations.

This is most useful if the global delete_mode is
set to ‘keep’, allowing a cleanup job (e.g. using
map/reduce) to later delete with mode ‘immediate’
or a time interval.

The parameter delete_mode is also exposed to http
and protocol buffer clients.