Fix dataloss bug caused by reusing causal histories
See http://github.com/riak_kv/issues/679 for a full history and
discussion. In summary: certain scenarios may lead to riak breaking a
strict invariant of version vectors, namely that each actor issue an
increasing, total order of events. When this invariant is broken some
new data item may be lost when some older item appears to be from a
later causal time. This fix adds per-key-epochs, so causal histories
are not re-used.

Examples:
----

Tombstone->doomstone
- Key is deleted
- Tombstone is written to 2 primaries and a fallback
- fallback is unavailable
- tombstone read-repaired over 3 primaries and reaped
- new key written
- fallback hands off
- old tombstone subsumes new key, data lost

Dataloss
- A replica becomes unavailable in a put coordinating node due to corruption or user error.
- The incoming object may or may not contain a vector clock.
- We get a not_found in the local put. We have lost the information corresponding to this vnode id.
- The vector clock created could have a stale number for this vnode
  actor id. Either 1 if no vclock or < than the real one if more
  writes have happened since.
- The vector clock is subsumed by the other replicas when sent out, so
  this write will lose and will eventually be read repaired completely
  out of existence.

There are others, but these are the two that this commit addresses.
See the kv679* tests in riak_test for more details.

How it Does it:
----

When a coordinating vnode updates a key it has no local clock entry
for, it creates a new vnode id for that key only, using its base
vnodeid and a monotonic per vnode epoch counter, this ensures that any
version vector from the past/future would conflict with the new one,
not overwrite it. We use a counter lease scheme, and async leasing to
keep the counter update durable and quick. The actor is per-key, so no
actor general actor explosion occurrs.

There is an off switch {riak_kv, [{use_epoch_counter, false}]}. The
default is `true`.

Not done:
----

There is another flavour of the issue, see private issue tracker, that
is not yet addressed, but can be in future with some extra clock
metadata.