Don't force a new epoch on old keys
use the bare vnodeid as long as possible, only start a new epoch
when needed. In a way the bare vnodeid is just epoch zero, so use it.

Also, consider the incoming counter when deciding about new epochs.
An incoming clock with the same actor+epoch but greater counter is
a hint that a byzantine failure occurred and a new epoch is needed