WIP integrate with new map/orswot
Add capability and env_var to control binary format of map/set

See basho/riak#667 (RIAK-1436) (RIAK-1482) for more details.

This PR adds a capability for the binary output of Maps and Sets in
riak. It also uses and app_env {riak_kv, crdts_mixed_versions} to
override the capability so that an operator may force the riak cluster
to use v1 binary format until all clusters in a multi-datacenter set up
are upgraded to riak-2.0.x>4.

There is a cost associated with running in mixed cluster mode. See
basho/riak_dt#111 for details. A v2 map written to v1 must fully recurse
the structure to convert to a format readable by v1.