Ignore warnings about prepare_index_query.
For some reason, Dialyzer believes that the caller to this function will
only be undefined or {_, _}, but it's unclear where this assumption is
coming from; the caller of this function comes in through a coverage
query through riak_core, so dialyzer is deriving this call from some
unknown place.

Ignore the warnings for the time being.