Do not crash riak node if vnode status file is unreadable
If for any reason the vnode status file is unreadable with
`file:consult/1` the vnode would start, fail to consult, crash until
the restart limit was hit and the node would fail to start/crash.

This change makes that more robust by simply throwing away a status file
that cannot be read, and using a new, empty vnode status. This has the
cost of a new vnodeid+epoch counter for the vnode, but is more robust.

This fix uses the code added to riak_core under the same branch name
to always make a fair attempt to read the vnode status file.

In erlang 17 the default file encoding was switched to utf8, prior
to that it was latin1. Some vnode status files written in r16 are
un-consultable in 17 simply for this reason. See riak_core for more
on that.

Some vnode files are un-consultable because they were written with
io_lib:format("~p.", [Term]). `io:format' docs say:

"Writes the data with standard syntax in the same way as ~w, but breaks
terms whose printed representation is longer than one line into many
lines and indents each line sensibly. It also tries to detect lists of
printable characters and to output these as strings."

Which is why this change also use `io_lib:format("~w.", [Term])` to
write vnode status files. This is backwards compatible so these status
files can be read on r16 and previous riaks in the case of downgrade.