DP-6 by default, check for existence of {Bucket, Key}
As get_api_entry_points gets the list of vnodes where a {Bucket, Key}
is stored from riak_core_apl:get_apl_ann (which just computes a chash
for that BKey to figure vnodes), we get the entry points for BKey
regardless of whether such BKey actually exists.  This may be
undesirable, so we now (optionally do not) call riak_client:get to
account for this case and report an empty list data are not there.

The newly added option to suppress this check is check_key_exist (same
option appears in RpbApiEpReq).