Fix crash on vnode start if status file is mixed 1 and 2
Reported by booking.com in kv1142: an upgrade to 2.1, followed by
a downgrade to 2.0, followed by an upgrade to 2.1 caused  them to
have mixed vnode_status files, where `counter` was present, but
the versin was `1`. A none-exhaustive case statement in vnode
status manager did not handle the case, and the vnode could not
start, leading to a node that would not start.

A simpler way to get into this mess is simply to turn on epochs,
turn them off, and then trun them on  again.

This fix deals with mixed vnode_status files