Add bucket property validation for last_write_wins xor dvv_enabled
https://bashoeng.atlassian.net/browse/RIAK-1743

If `last_write_wins` is true then `dvv_enabled` cannot also be true. Add
bucket property validation to ensure that this is the case. The bucket
property validation applies to create and update on any buckets or
bucket types. Since this validation applies to a combination of two
separate properties in both the create and update scenarios, a new
`validate_post_merge` function was added, which allows for validation on
properties after they have been fully resolved and merged.