don't reverse order of records when collecting chunks for fetch
Because the data were coming in reverse order (note the way new
elements are prepended in riak_pb_ts_codec:row_for/2), the records
were sent to leveldb in reverse order.  However, another reversal was
done on fetch, the cumulative effect being the eventually correct
order as presented to the client.

As leveldb has optimisations in place relying on the TS data in the
un-reversed order, the double reversal is a bug.  This commit is a
fix.