Split 'assert' into mupltiple 'assertEqual'.
This should make future readings easier, since the msgpack test is now
similar in style to the test surrounding it. Also, this will give a more
exact failure point then using '==' and boolean comparisons.

A direct pattern match (Obj = DecodeResult) would have acheived a similar
goal, but the prefered style is to use ?assertEqual to get verbose output
on failure.