avoid riak_core_claimant:bucket_type_status, rely on _core:get_bucket
The call to claimant to determine a table's bucket type status is an
expensive operation.  It was introduced in a1c1e2e6, with the purpose to
detect and report the error condition where a TS operation is attempted on
a non-activated bucket type, but found to cause serious performance
degradation.

This error condition is incidentally also detected by
riak_core_bucket:get_bucket/1, a call used to get bucket properties in
many code paths throughout riak_kv. If called on a non-activated bucket,
it returns {error, no_type}. The proposed patch uses this to detect and
report this type of failure while avoiding the costly call to
riak_core_claimant:bucket_type_status.