detect and deal with {error,no_type} (table not acivated case)
Complementary to 9b62335 by atill

The call to claimant to determine a table's bucket type status is an
expensive operation.  It was introduced in a1c1e2e6, with the purpose to
detect and report the error condition where a TS operation is attempted on
a non-activated bucket type, but found to cause serious performance
degradation. Andy's commit removed the expensive call; this commit captures
and reports the error condition in various places in
riak_kv_{wm,pb}_timeseries and riak_kv_ts_util.