relax DDL check in is_type_compiled
The check for type availability uses a match on DDL bodies of argument
DDL and that previously compiled and stored in metadata. That check is
not strictly necessary, but more importantly, it is bound to fail when
DDLs are produced from the same schema but in different #ddl_vX{} versions.

Relaxing this check will allow post-1.2 nodes to see DDLs compiled by
different ddl record versions, and thus do not be a blocker in
upgrade/downgrade scenarios.