fix list_keys for short keys
It appears short keys, as in

  CREATE TABLE t (ts TIMESTAMP NOT NULL, PRIMARY KEY((ts),ts));

don't get sext-encoded, and are instead sent to the calling process
of riak_client:stream_list_keys as tuples, causing the resource streaming
fun to silently choke on sext:decode(Tuple). As a result, the http client
expecting chunked response, receives nothing.

This is probably a bug on the side of riak_kv, which we work around by
explicitly not attempting a sext:decode on tuples.