Remove unused riak_kv_put_fsm start/3 function
This was added way back in the 1.4 days, but its only usage was removed
shortly thereafter since it introduced an API incompatibility that broke
mixed clusters. The implementation has been sitting around unused since
then, and it should be fine to remove it (just to be safe, I confirmed
that no versions of Riak since before 2.0 have used this, so even if it
was ever used in a released version of Riak, it was over 2 major
versions in the past so we don't need to support it any longer).