Fix get_fsm_eqc fail from new stats path being hit
Since the non-sidejob stats path has been removed, the code we're
testing no longer uses exometer data to determine active FSM counts, and
instead queries sidejob. This is the same way the code works in
production when sidejob is enabled, but the get_fsm_eqc test was using
the non-sidejob path, so it was faking the usage data by inserting
values into exometer. This fixes the test by removing the fake exometer
data and instead mocking the corresponding sidejob calls using meck.