new convenience fun riak_kv_ts_svc:lk_to_pk/3
to use in key conversion function for proper coverage plan calculation
instead of ts_util:row_to_key, in list_keys code path. Specifically,
this is 01725aa15c done right.

As its argument, the old key conversion function (row_to_key) used to take
the LK values while it logically treated the argument as a list of *all*
values of a TS record. For the (common) case where key fields all appear
at the beginning of DDL, and they do so in the order they are defined,
row_to_key happens to work. It will break, however, when that condition
does not hold.

To fix this, we now use a newly written function lk_to_pk, which does
what its name suggests, and which is a pared down version of make_ts_keys.