Added logic to hash only vclock for object hashes. Added logging to catch pre 2.1 vclocks which require a rewrite to be safe
Added riak_kv boolean environment variable `hash_only_vclock` which
enables/disabled the vclock hashing logic. Additionally, when enabled,
we will no longer do the expensive compare/logging on read-repair to
catch old pre2.1 vclocks who require a rewrite.

Additional logic must be added to `yz_kv.erl` so the hashes in the
entropy_data element of the Solr documents match what is stored in KV
AAE. Upgrading the hash logic requires clearing all YZ documents and
re-indexing everything. This is because we store the obj hash in the
Solr document. This will hopefully change in the future to allow for
easier AAE changes.