Don't update trees before deleting them from disk.
Allow the case where we're about to destroy the hashtree to simply close,
without going through a (sometimes very lengthy) update process first. Only
used from `destroy_trees`. This condition was caught by a 4-day performance
test, and is not generally reproducible without having tons of data, on
the order of 10s of millions of key/value pairs per partition.