Make fold read new build limit after each sleep in `maybe_throttle_build`.
Before, the build limits (size and wait time) could only be changed
before a build started. Once it did, you were stuck with those settings.
If the vnode was large, you may want to decrease the wait or increase
the size, but we wouldn't pick it up until the rebuild was over.