Refactor request records into request module
Remove definition and usage of the request macros, for example:
* ?KV_PUT_REQ
* ?KV_GET_REQ
* ?KV_LISTKEYS_REQ

Replace with factories and getters in riak_kv_requests. Within
riak_kv_vnode, we replace pattern matching on the record with
dispatch to functions that match on a request_type getter.

Refactor use of case AsyncFolding to use existing function

Less fun: remove unnecessary function application

Inline module name so it is visible to xref

Extract `fold_type_for_query` function

Also, delete `prepare_index_query` function, as it was nothing more than
an identity function used in exactly one place, and had a bad type spec
to boot. While we were at it, deleted the lines in the
`dialyzer.ignore-warnings` file that referenced the now-deleted
`prepare_index_query` function.