Add test for sweeper run_interval functionality
The run_interval field in the #sweep_participant{} record is poorly
named. When set it will cause the sweep_participant to be rescheduled
after run_interval seconds have elapsed after the last sweep.

Note that we are receiving the {ok, successfull_sweep, ...}  message,
sent by the mocked successfull_sweep callback function twice because the
scheduler implementation distinguishes between _never run_ sweeps, which
have the highest priority, and those to be scheduled periodically. It's
the second receive_msg calls that tests scheduling via run_interval.

Note also the second recieve_msg calls using 2500 msecs as the wait
timeout instead of the default. It seems to be the smallest value for
which the test completes successfully.