Fix moved riak_kv_sweeper_SUITE to itest directory
By convention eunit tries to auto discover tests in the test directory
and thinks all functions of arity 0 suffixed with _test. Moving the
common tests to a different directory avoids accidentaly discovery. Note
common testcases are functions with arity 1 but there can be an optional
info function with the same of 0 arity that returns properties of the
test cases. The problem was discovered because some testscases were
incorrectly created as 0 arity functions - FIXED.