Change move sweeper callback functions to their own module
It is confusing when reading the riak_kv_sweeper code to have callback
functions that execute in a riak_kv_vnode owner process mixed together
with those from the riak_kv_sweeper gen_server code. For example, one
can not know which process self() just by looking at the function in
isolation or why we are using gen_server:cast(...) in some of the
function.

The module with the callbacks is named riak_kv_sweeper_fold for lack of
a better name.