Add test case to remove sweep participant when scheduler is running
remove_sweep_participant will try to stop all running sweeps for the
given participant. To test this we need a new visit_function to ensure
we are in a running sweep. The new visit_function synchronizes with a
controller process, visit_function_controller, that synchronizes with
the test case; so the test case receives a message that the sweep on an
index is now waiting and can remove the sweep participant and let the
sweep continue.

The visit_function_controller is required because the scheduler picks an
index to sweep at random. So we can not wait in a fixed order of
indices. An alternate to the visit_function_controller process
implementation is a callback to visit_function where one can remove the
sweep participant.

This test discovered a bug we did not handle throw correctly when
folding over the keys - FIXED.