Add tests cases for obj_size throttle strategy
Throttling has no observable effect except for waiting, so to test
without resorting to mocking, a variable tracking the total throttle
wait time is added. The obj_size throttling strategy depends on two
parameters

 - amount of data read
 - number mutations to an object during the sweep

When counting amount of data read and overhead of 36 bytes headers in
the serialized riak object is included.

When counting the number of mutations for extra throttling there there
is a hard coded value of 100 mutated objects that triggers an extra
throttle.

The tests found a bug in the with incorrect precedence of + and
rem when counting mutated objects - FIXED.