Fix put FSM race condition (#1526) (#1527)
Cannot rely on a gen_fsm timeout event being sent even when the timeout
is set to zero. In particular, another message or event can preempt and
unset the timeout if the handler of that event does not reset it. Rather
than trying to save and restore the timeout state, manually send a
timeout event using gen_fsm:send_event.