Change move logic for sweep request to scheduler
Refactor sweep_request(...). The logic for scheduling a request was
spread on two paths for ticks and requests making it difficult
understand and test and fragile. For example, the concurrency limit
check was on two different paths in the code. Now all logic for
scheduling, either a tick or an explicit request, are in one functions
with EQC properties for it.