Fix race between capabilities registration and AAE startup.
During testing, it was found to be possible for a hashtree to start
_before_ the `{riak_kv, object_hash}` capability was registered, causing
the wrong version to be chosen when the tree started up. This can cause
old trees to be loaded, ignoring new trees that have more recent data,
and the wrong hash version to be used, which can cause unnecessary
repairs.

Moving the registration of the anti-entropy-related capabilities
to `riak_kv_entropy_manager` and changing the order of startup in
`riak_kv_sup` to make sure the entropy manager is started before
the vnodes start ensures that the capability is registered and
negotiated before hashtrees try to start.