Rename maybe_initiate_sweeps function
There's no "maybe" about it. It will always update the sweep records in
the sweeper state. "Initiate" is also a bad choice of words: we're not
starting any sweeps, as the word initiate implies. We're just updating
the sweeps records in our state which tell us which sweeps we're
responsible for running.