make sure follow-up queries match original one
This is to prevent a follow-up query to be directed to an existing query
buffer if the user supplies a non-matching follow-up. It is an error.

Note that as of now, this code is dead (it relies on a real client-supplied
qbuf_id, to be implemented in 1.6). Every query gets a new, unique qbuf_id.