Add exometer tracking for keys and bytes swept
Using an exometer sprial probe we report the keys swept and the bytes
swept per index. The spiral will sum the reported values for an
interval giving us a the rate of processing keys and bytes per
index. The effect of sweep concurrency and throttling can also be
visualized by plotting these metrics.