Add ignore warnings for use of `consistent` in field.
Before GPB, record fields were not typed. Now that they are (which is a
good thing) we have some abuses of the record fields internal to
`riak_kv_pb_object.erl` that we'll have to work out later. For now,
ignore the warnings so others can get back to work.