Silently take blob data over HTTP as base64
Chris Mancini identified a key issue: JSON does *not* like binary
data. Since JSON is our sole interface for inserting and retrieving
records via HTTP, that's a bit of a problem.

This set of changes allows HTTP clients such as our PHP library to use
the industry-standard base64 encoding to upload and retrieve blob data
for insert, query, get, and list_keys.

Yet to be done: supporting blob as primary key values for get, delete,
list_keys. (Data returned from list_keys is supported).

Also must consider branching the logic to allow 0x strings to be sent
instead of base64 to maintain consistency with the SQL command
line. (This will be the only way to do SQL queries with blob as key
values). Outstanding question: which format should we use for output?