qry_worker: kill outstanding FSMs on cancelling query
When a query is cancelled (because of div_by_zero or if its projected size
is found to exceed configured limit), there may be index FSMs on vnodes
still busy collecting chunks.  Those chunks will be sent to the
coordinator when they are ready, where the query worker instances to
receive them will be gone.

It makes sense to terminate the FSMs early, once it is known the data
they are collecting will be discarded.