qry_buffers: in-mem ORDER BY
The underlying idea is to keep the accumulated data in memory until memory
shortage occurs in the qry_buffers gen_server process. The two costliest
operations, eleveldb:open and :put (as well as sext en-/decoding) are thus
deferred, and for short queries and under light load, may never actually
take place.

The heuristic to determine when to start dumping accumulated query results
to leveldb is, in this initial implementation, simply a check of

  (heap_size - stack_size) < HeapSizeLimit

with the limit set in riak.conf by parameter
riak_kv.timeseries_query_buffers_inmem_max (10M by default).