Bring index_hashtree fixes from 2.2 to 2.0 (#1574)
* Don't update trees before deleting them from disk.

Allow the case where we're about to destroy the hashtree to simply
close,
without going through a (sometimes very lengthy) update process first.
Only
used from `destroy_trees`. This condition was caught by a 4-day
performance
test, and is not generally reproducible without having tons of data, on
the order of 10s of millions of key/value pairs per partition.
* Handle potential errors in vnode fold request call when rebuilding.

`fold_fun`'s call to `riak_core_master:sync_command` ignored the result
of the call. It is possible for `sync_command` to return {error,
overload}
if the vnode is overloaded. Rather than ignore the result, make sure to
handle non-success results by failing the tree build.
* Make fold read new build limit after each sleep in `maybe_throttle_build`.

Before, the build limits (size and wait time) could only be changed
before a build started. Once it did, you were stuck with those settings.
If the vnode was large, you may want to decrease the wait or increase
the size, but we wouldn't pick it up until the rebuild was over.