Remove code for riak-admin reformat-indexes
This was only needed for upgrading from versions of Riak prior to 1.3.1.
We're well past the point where we no longer need to support this, so
I'm removing the associated code.

This commit removes the code for actively reformatting indexes, but
leaves some of the old code in riak_kv_eleveldb_backend for handling
index upgrades and downgrades. That code will be a bit trickier to
remove, since we'll want to leave some level of protection in place just
in case somebody does try to skip a bunch of versions when upgrading.
Will attempt to tackle that in the next commit.

Conflicts:
	dialyzer.ignore-warnings