deferred eleveldb ops for query buffers (in-mem ORDER BY/LIMIT) (#1587)
* qry_buffers: prepare keyed data before calling ldb:put
* qry_buffers: in-mem ORDER BY

The underlying idea is to keep the accumulated data in memory until memory
shortage occurs in the qry_buffers gen_server process. The two costliest
operations, eleveldb:open and :put (as well as sext en-/decoding) are thus
deferred, and for short queries and under light load, may never actually
take place.

The heuristic to determine when to start dumping accumulated query results
to leveldb is, in this initial implementation, simply a check of

  (heap_size - stack_size) < HeapSizeLimit

with the limit set in riak.conf by parameter
riak_kv.timeseries_query_buffers_inmem_max (10M by default).
* Merge branch 'riak_ts-develop' into feature-az-inmem_qbufs
* schema: use bytesize type for timeseries.qbuf_inmem_max
* qry_buffers: unoverload #state.ldb_ref
* qry_buffers: clarify the use of ChunkId in artificial key construction
* qry_buffers: rename some identifiers
* qry_buffers: allow incomplete buffers to be deleted
* qry_buffers: don't try to qry_buffers_ldb:delete_table(LdbRef = undefined)
* qry_buffers: rename more identifiers
* fixup for 75b04e7033
* qry_buffers: split big function add_chunk in two
* qry_buffers: add a spec for enkey/4
* qry_buffers: comment touchup
* qry_buffers: better explain when we switch to writing to backend
* Add support for GPB protocol buffer library
* Merge branch 'bch-add-gpb-support' into feature-az-inmem_qbufs
* Merge branch 'riak_ts-integration' into feature-az-inmem_qbufs

# Conflicts:
#	protogen.config
* qry_buffers: simplify, fix make_sorted_keys/3 wrt NULLS FIRST/LAST