Address RIAK-2976
Add logic to skip waiting for all exchanges to complete if other nodes
in the cluster have already flipped either their `version` or
`pending_version` variable. Only a single node reporting back an
upgraded variable is good enough to go ahead and start the upgrade.
This fixes a bug where a joining nodes hash trees would never upgrade
if the join occurred during of after the initial hash tree upgrade for
the cluster.