Handle empty `trees_version` in upgrade path
In some specific cases, we could hit the upgrade tick before any hash
trees have registered with the entropy manager. Make sure we don’t
assume the upgrade has finished and instead skip this tick and check
again on the next tick.

Found using various tests which wait on AAE repairs. Discovered in rare
cases when an upgrade check was triggered immediately after the node
joined the cluster but before the plan is committed. This resulted in a
`version` of `0` but all the `trees_versions` would still be `legacy`.

Move pending_version() API to use ETS instead of calling into the
entropy_manager process for the current state. The manager is used for
many essential functions and should not be blocked waiting on a
response from another manager.