Add basic failing eqc for kv679 bugs
In this test all writes are coordinated by the head of the preflist,
pfhead, and replicated to at least one other node chosen from 2 further
primaries and 3 fallbacks. Every write is there for w=2 and acked on
pfhead one other node. Every now and again one node may replicate to
another. Every now and again pfhead may "forget" it's local values (this
is like a disk error or partially successful delete.) In this commit the
test catches the lost acked write bug pretty quickly. Next up is to add
the vnode counter and see if the test can catch the case as yet unfixed.

Also added a convenience function to riak_object.