Add epochs to non-coordinating writes
This commit adds a no-op/no-dot VV increment with a new epoch actor to
an object if it is needed. If the incoming object contains an entry
for the vnode greater than the local one then that indicates that the
local vnode has suffered amnesia and needs to ensure it uses a new
actor id next time it coordinates a write on this key. The simplest
way to do that is add a VV entry of Actor:Epoch+1 with a counter of 1.
Note the new function on riak_object as this increment must _NEVER_
create a dot.