Make sweeper config naming consistent
We have one setting name prefixed with "sweeper." and some other ones
starting with "riak_kv.sweep."

All the sweeper options should start with the same prefix. Additionally,
since "Riak KV" means "Riak, the KV store database" to the general
public, we shouldn't be prefixing settings with "riak_kv" (especially
since we could merge the sweeper over to Riak TS at some point ;-)

Using the "sweeper." prefix everywhere is succinct and helpful, since
Cuttlefish is meant to be user-friendly, and customers will think of the
Sweeper as its own distinct component. Which application it happens to
be located in is probably irrelevant to them, hence my decision to kill
the "riak_kv." prefix.